Create a Point_set_processing_3 folder to move internal
