+++ Point_set_processing_3/include/CGAL/Point_set_processing_3/internal/Rich_grid.h
